requirements/production.txt
-r base.txt

dj-database-url
gunicorn==22.0.0
psycopg2-binary==2.9.9
whitenoise==6.5.0